refactoring tut11
