+++ VulkanTest/VulkanTest.vcxproj
@@ -166,16 +166,20 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="first_app.cpp" />
+    <ClCompile Include="kc_custom.cpp" />
     <ClCompile Include="lve_game_object.cpp" />
     <ClCompile Include="lve_model.cpp" />
+    <ClCompile Include="lve_renderer.cpp" />
     <ClCompile Include="lve_window.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="lve_pipeline.cpp" />
     <ClCompile Include="lve_device.cpp" />
     <ClCompile Include="lve_swap_chain.cpp" />
+    <ClCompile Include="simple_render_system.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="first_app.hpp" />
+    <ClInclude Include="kc_custom.hpp" />
     <ClInclude Include="lve_game_object.hpp" />
     <ClInclude Include="lve_model.hpp" />
     <ClInclude Include="lve_renderer.hpp" />
@@ -183,6 +187,7 @@
     <ClInclude Include="lve_pipeline.hpp" />
     <ClInclude Include="lve_device.hpp" />
     <ClInclude Include="lve_swap_chain.hpp" />
+    <ClInclude Include="simple_render_system.hpp" />
   </ItemGroup>
   <ItemGroup>
     <None Include="compile.bat" />
