+++ VulkanTest/first_app.cpp
@@ -1,5 +1,8 @@
 #include "first_app.hpp"
 
+#include "kc_custom.hpp"
+#include "simple_render_system.hpp"
+
 // libs
 #define GLM_FORCE_RADIANS
 #define GLM_FORCE_DEPTH_ZERO_TO_ONE
@@ -12,73 +15,26 @@
 #include <memory>
 #include <stdexcept>
 
-namespace {
-// temp code
-struct SimplePushConstantData {
-  glm::mat2 transform{1.0f};
-  glm::vec2 offset;
-  alignas(16) glm::vec3 color;
-};
-std::vector<lve::LveModel::Vertex> generate_sierpinski(
-    int level, const std::vector<lve::LveModel::Vertex>& current_triangles) {
-  std::vector<lve::LveModel::Vertex> return_triangles;
-  if (level == 0) {
-    return_triangles = current_triangles;
-  } else {
-    std::vector<lve::LveModel::Vertex> prev_triangles;
-    prev_triangles = generate_sierpinski(level - 1, current_triangles);
-
-    int num_triangles = static_cast<int>(prev_triangles.size()) / 3;
-    for (int i = 0; i < num_triangles; i++) {
-      for (int j = 0; j < 3; j++) {
-        int prev_j = (j + 3 - 1) % 3;
-        int next_j = (j + 1) % 3;
-        const lve::LveModel::Vertex& current_vertex = prev_triangles[3 * i + j];
-        const lve::LveModel::Vertex& prev_vertex =
-            prev_triangles[3 * i + prev_j];
-        const lve::LveModel::Vertex& next_vertex =
-            prev_triangles[3 * i + next_j];
-
-        glm::vec2 p1 = current_vertex.position;
-        glm::vec2 p2 =
-            glm::mix(current_vertex.position, next_vertex.position, 0.5);
-        glm::vec2 p3 =
-            glm::mix(current_vertex.position, prev_vertex.position, 0.5);
-
-        glm::vec3 c1 = current_vertex.color;
-        glm::vec3 c2 = glm::mix(current_vertex.color, next_vertex.color, 0.5);
-        glm::vec3 c3 = glm::mix(current_vertex.color, prev_vertex.color, 0.5);
-
-        return_triangles.push_back({p1, c1});
-        return_triangles.push_back({p2, c2});
-        return_triangles.push_back({p3, c3});
-      }
-    }
-  }
-  return return_triangles;
-}
-}  // namespace
-
 namespace lve {
 
-FirstApp::FirstApp() {
-  loadGameObjects();
-  createPipelineLayout();
-  recreateSwapChain();
-  createCommandBuffers();
-}
-FirstApp::~FirstApp() {
-  vkDestroyPipelineLayout(lveDevice.device(), pipelineLayout, nullptr);
-}
+FirstApp::FirstApp() { loadGameObjects(); }
+FirstApp::~FirstApp() {}
 
 void FirstApp::run() {
-  {
-    while (!lveWindow.shouldClose()) {
-      glfwPollEvents();
-      drawFrame();
+  SimpleRenderSystem simpleRenderSystem{lveDevice,
+                                        lveRenderer.getSwapChainRenderPass()};
+  while (!lveWindow.shouldClose()) {
+    glfwPollEvents();
+    if (auto commandBuffer = lveRenderer.beginFrame()) {
+      // NOTE: separate frame and renderpass, since we need to control
+      // multiple render passes.
+      lveRenderer.beginSwapChainRenderPass(commandBuffer);
+      simpleRenderSystem.renderGameObjects(commandBuffer, gameObjects);
+      lveRenderer.endSwapChainRenderPass(commandBuffer);
+      lveRenderer.endFrame();
     }
-    vkDeviceWaitIdle(lveDevice.device());
   }
+  vkDeviceWaitIdle(lveDevice.device());
 }
 
 void FirstApp::loadGameObjects() {
@@ -86,7 +42,7 @@ void FirstApp::loadGameObjects() {
                                          {{0.5f, 0.3f}, {0.0f, 1.0f, 0.0f}},
                                          {{-0.5f, 0.3f}, {0.0f, 0.0f, 1.0f}}};
   std::vector<LveModel::Vertex> sierpinski_vertices =
-      ::generate_sierpinski(2, vertices);
+      kc_custom::generate_sierpinski(2, vertices);
   // NOTE(kc): a model can be shared by multiple game objects.
   auto lveModel = std::make_shared<LveModel>(lveDevice, sierpinski_vertices);
 
@@ -108,182 +64,6 @@ void FirstApp::loadGameObjects() {
     triangle.transform2d.rotation = i * .025f * glm::pi<float>();
 
     gameObjects.push_back(std::move(triangle));
-  }
-}
-
-void FirstApp::createPipelineLayout() {
-  VkPushConstantRange pushConstantRange{};
-  pushConstantRange.stageFlags =
-      VK_SHADER_STAGE_VERTEX_BIT | VK_SHADER_STAGE_FRAGMENT_BIT;
-  pushConstantRange.offset = 0;
-  pushConstantRange.size = sizeof(SimplePushConstantData);
-
-  VkPipelineLayoutCreateInfo pipelineLayoutInfo{};
-  pipelineLayoutInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-  pipelineLayoutInfo.setLayoutCount = 0;
-  pipelineLayoutInfo.pSetLayouts = nullptr;
-  pipelineLayoutInfo.pushConstantRangeCount = 1;
-  pipelineLayoutInfo.pPushConstantRanges = &pushConstantRange;
-  if (vkCreatePipelineLayout(lveDevice.device(), &pipelineLayoutInfo, nullptr,
-                             &pipelineLayout) != VK_SUCCESS) {
-    throw std::runtime_error("failed to create pipeline layout!");
-  }
-}
-void FirstApp::createPipeline() {
-  assert(lveSwapChain != nullptr &&
-         "Cannot create pipeline before swap chain.");
-  assert(pipelineLayout != nullptr &&
-         "Cannot create pipeline before pipeline layout.");
-
-  PipelineConfigInfo pipelineConfig{};
-  LvePipeline::defaultPipelineConfigInfo(pipelineConfig);
-  pipelineConfig.renderPass = lveSwapChain->getRenderPass();
-  pipelineConfig.pipelineLayout = pipelineLayout;
-  lvePipeline = std::make_unique<LvePipeline>(
-      lveDevice, "./shaders/simple_shader.vert.spv",
-      "./shaders/simple_shader.frag.spv", pipelineConfig);
-}
-
-void FirstApp::recreateSwapChain() {
-  auto extent = lveWindow.getExtent();
-  // ex minimize
-  while (extent.width == 0 || extent.height == 0) {
-    extent = lveWindow.getExtent();
-    glfwWaitEvents();
-  }
-  vkDeviceWaitIdle(lveDevice.device());
-  if (lveSwapChain == nullptr) {
-    lveSwapChain = std::make_unique<LveSwapChain>(lveDevice, extent);
-  } else {
-    lveSwapChain = std::make_unique<LveSwapChain>(lveDevice, extent,
-                                                  std::move(lveSwapChain));
-    if (lveSwapChain->imageCount() != commandBuffers.size()) {
-      freeCommandBuffers();
-      createCommandBuffers();
-    }
-  }
-  // some systems two swapchain cant coexist on the same window.
-  // lveSwapChain = nullptr;
-  createPipeline();
-}
-void FirstApp::createCommandBuffers() {
-  commandBuffers.resize(lveSwapChain->imageCount());
-
-  VkCommandBufferAllocateInfo allocInfo{};
-  allocInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_ALLOCATE_INFO;
-  allocInfo.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
-  allocInfo.commandPool = lveDevice.getCommandPool();
-  allocInfo.commandBufferCount = static_cast<uint32_t>(commandBuffers.size());
-
-  if (vkAllocateCommandBuffers(lveDevice.device(), &allocInfo,
-                               commandBuffers.data()) != VK_SUCCESS) {
-    throw std::runtime_error("failed to allocate command buffers!");
-  }
-}
-
-void FirstApp::freeCommandBuffers() {
-  vkFreeCommandBuffers(lveDevice.device(), lveDevice.getCommandPool(),
-                       static_cast<uint32_t>(commandBuffers.size()),
-                       commandBuffers.data());
-  commandBuffers.clear();
-}
-
-void FirstApp::recordCommandBuffer(int imageIndex) {
-  VkCommandBufferBeginInfo beginInfo{};
-  beginInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO;
-  if (vkBeginCommandBuffer(commandBuffers[imageIndex], &beginInfo) !=
-      VK_SUCCESS) {
-    throw std::runtime_error("failed to begin recording command buffer!");
-  }
-
-  VkRenderPassBeginInfo renderPassInfo{};
-  renderPassInfo.sType = VK_STRUCTURE_TYPE_RENDER_PASS_BEGIN_INFO;
-  renderPassInfo.renderPass = lveSwapChain->getRenderPass();
-  renderPassInfo.framebuffer = lveSwapChain->getFrameBuffer(imageIndex);
-
-  renderPassInfo.renderArea.offset = {0, 0};
-  renderPassInfo.renderArea.extent = lveSwapChain->getSwapChainExtent();
-
-  std::array<VkClearValue, 2> clearValues{};
-  clearValues[0].color = {0.01f, 0.01f, 0.01f, 1.0f};
-  clearValues[1].depthStencil = {1.0f, 0};
-  renderPassInfo.clearValueCount = static_cast<uint32_t>(clearValues.size());
-  renderPassInfo.pClearValues = clearValues.data();
-
-  vkCmdBeginRenderPass(commandBuffers[imageIndex], &renderPassInfo,
-                       VK_SUBPASS_CONTENTS_INLINE);
-  VkViewport viewport{};
-  viewport.x = 0.0f;
-  viewport.y = 0.0f;
-  viewport.width = static_cast<float>(lveSwapChain->getSwapChainExtent().width);
-  viewport.height =
-      static_cast<float>(lveSwapChain->getSwapChainExtent().height);
-  viewport.minDepth = 0.0f;
-  viewport.maxDepth = 1.0f;
-  VkRect2D scissor{{0, 0}, lveSwapChain->getSwapChainExtent()};
-  vkCmdSetViewport(commandBuffers[imageIndex], 0, 1, &viewport);
-  vkCmdSetScissor(commandBuffers[imageIndex], 0, 1, &scissor);
-
-  renderGameObjects(commandBuffers[imageIndex]);
-
-  vkCmdEndRenderPass(commandBuffers[imageIndex]);
-  if (vkEndCommandBuffer(commandBuffers[imageIndex]) != VK_SUCCESS) {
-    throw std::runtime_error("failed to record command buffer!");
-  }
-}
-
-void FirstApp::renderGameObjects(VkCommandBuffer commandBuffer) {
-  // update
-  int i = 0;
-  for (auto& obj : gameObjects) {
-    i++;
-    obj.transform2d.rotation = glm::mod<float>(
-        obj.transform2d.rotation + 0.00001f * i, glm::two_pi<float>());
-  }
-  // render
-  lvePipeline->bind(commandBuffer);
-  for (auto& obj : gameObjects) {
-    SimplePushConstantData push{};
-    push.offset = obj.transform2d.translation;
-    push.color = obj.color;
-    push.transform = obj.transform2d.mat2();
-
-    vkCmdPushConstants(
-        commandBuffer, pipelineLayout,
-        VK_SHADER_STAGE_VERTEX_BIT | VK_SHADER_STAGE_FRAGMENT_BIT, 0,
-        sizeof(SimplePushConstantData), &push);
-    obj.model->bind(commandBuffer);
-    obj.model->draw(commandBuffer);
-  }
-}
-
-void FirstApp::drawFrame() {
-  uint32_t imageIndex;
-  auto result = lveSwapChain->acquireNextImage(&imageIndex);
-
-  // window resized
-  if (result == VK_ERROR_OUT_OF_DATE_KHR) {
-    recreateSwapChain();
-    return;
-  }
-
-  if (result != VK_SUCCESS && result != VK_SUBOPTIMAL_KHR) {
-    throw std::runtime_error("failed to acquire swap chain image!");
-  }
-
-  recordCommandBuffer(imageIndex);
-  result = lveSwapChain->submitCommandBuffers(&commandBuffers[imageIndex],
-                                              &imageIndex);
-  // suboptimal : no longer matches the surface properties exactly, but can
-  // still be used.
-  if (result == VK_ERROR_OUT_OF_DATE_KHR || result == VK_SUBOPTIMAL_KHR ||
-      lveWindow.wasWindowResized()) {
-    lveWindow.resetWindowResizedFlag();
-    recreateSwapChain();
-    return;
-  }
-  if (result != VK_SUCCESS) {
-    throw std::runtime_error("failed to present swap chain image!");
   }
 }
