+++ VulkanTest/first_app.hpp
@@ -2,9 +2,7 @@
 
 #include "lve_device.hpp"
 #include "lve_game_object.hpp"
-#include "lve_model.hpp"
-#include "lve_pipeline.hpp"
-#include "lve_swap_chain.hpp"
+#include "lve_renderer.hpp"
 #include "lve_window.hpp"
 
 // std
@@ -26,20 +24,10 @@ class FirstApp {
 
  private:
   void loadGameObjects();
-  void createPipelineLayout();
-  void createPipeline();
-  void createCommandBuffers();
-  void freeCommandBuffers();
-  void drawFrame();
-  void recreateSwapChain();
-  void recordCommandBuffer(int imageIndex);
-  void renderGameObjects(VkCommandBuffer commandBuffer);
 
   LveWindow lveWindow{WIDTH, HEIGHT, "Hello Vulkan! ckc!"};
   LveDevice lveDevice{lveWindow};
-  std::unique_ptr<LveSwapChain> lveSwapChain;
-  std::unique_ptr<LvePipeline> lvePipeline;
-  VkPipelineLayout pipelineLayout;
+  LveRenderer lveRenderer{lveWindow, lveDevice};
   std::vector<VkCommandBuffer> commandBuffers;
   std::vector<LveGameObject> gameObjects;
 };
