+++ VulkanTest/lve_renderer.hpp
@@ -26,9 +26,14 @@ class LveRenderer {
   VkCommandBuffer getCurrentCommandBuffer() const {
     assert(isFrameStarted &&
            "Cannot get command buffer when frame not in progress.");
-    return commandBuffers[currentImageIndex];
+    return commandBuffers[currentFrameIndex];
   }
 
+  int getFrameIndex() const {
+    assert(isFrameStarted &&
+           "Cannot get frame index when frame not in progress.");
+    return currentFrameIndex;
+  }
   VkCommandBuffer beginFrame();
   void endFrame();
   void beginSwapChainRenderPass(VkCommandBuffer commandBuffer);
@@ -45,6 +50,7 @@ class LveRenderer {
   std::vector<VkCommandBuffer> commandBuffers;
 
   uint32_t currentImageIndex;
+  int currentFrameIndex;
   bool isFrameStarted;
 };
 }  // namespace lve
