+++ VulkanTest/lve_swap_chain.hpp
@@ -46,6 +46,11 @@ class LveSwapChain {
   VkResult submitCommandBuffers(const VkCommandBuffer *buffers,
                                 uint32_t *imageIndex);
 
+  bool compareSwapFormats(const LveSwapChain &swapChain) const {
+    return swapChain.swapChainImageFormat == swapChainImageFormat &&
+           swapChain.swapChainDepthFormat == swapChainDepthFormat;
+  }
+
  private:
   void init();
   void createSwapChain();
@@ -63,6 +68,7 @@ class LveSwapChain {
   VkExtent2D chooseSwapExtent(const VkSurfaceCapabilitiesKHR &capabilities);
 
   VkFormat swapChainImageFormat;
+  VkFormat swapChainDepthFormat;
   VkExtent2D swapChainExtent;
 
   std::vector<VkFramebuffer> swapChainFramebuffers;
